move ./Table to ./components/Table
